perf(crawlfilee): update file schema and improve perf upload file

+++ packages/server/src/module/files/services/files.service.ts
@@ -1,45 +1,59 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { S3 } from 'aws-sdk';
 import { ConfigService } from '@nestjs/config';
-import { v4 as uuid } from 'uuid';
 import { FileRepository } from '../repositories/file.repository';
+import { left } from '@/shared/core/Result';
+import { FileNotFoundError } from '@/shared/core/AppError';
 
 @Injectable()
 export class FilesService {
+  private readonly logger = new Logger(FilesService.name);
   constructor(
     private readonly configService: ConfigService,
     private fileRepository: FileRepository,
   ) {}
 
   async uploadFile(dataBuffer: Buffer, filename: string) {
+    this.logger.log(`Start uploading files into s3 bucket`);
     const s3 = new S3();
+
+    const createFile = await this.fileRepository.create({
+      name: filename,
+    });
+
     const uploadResult = await s3
       .upload({
         Bucket: this.configService.get('AWS_PUBLIC_BUCKET_NAME'),
         Body: dataBuffer,
-        Key: `${uuid()}-${filename}`,
+        Key: `${createFile._id}`,
         ACL: 'public-read',
       })
       .promise();
 
-    const createFile = await this.fileRepository.create({
-      key: uploadResult.Key,
+    const updateFile = await this.fileRepository.update(createFile?._id, {
       name: filename,
       url: uploadResult.Location,
     });
 
-    return createFile;
+    return updateFile;
   }
 
   async deleteFile(fileId: string) {
+    this.logger.log(`Start deleting file by fileId`);
     const file = await this.fileRepository.findById(fileId);
+
+    if (!file) {
+      return left(new FileNotFoundError());
+    }
+
     const s3 = new S3();
     await s3
       .deleteObject({
         Bucket: this.configService.get('AWS_PUBLIC_BUCKET_NAME'),
-        Key: file.key,
+        Key: file._id,
       })
       .promise();
-    await this.fileRepository.delete(fileId);
+
+    return this.fileRepository.delete(fileId);
   }
 }
